Add Filter.Tests.Common project
Centralize the logic for setting up a test SQL database and
Elasticsearch cluster/index.

Also, the glue code needed to figure out which SQL/ES instance
to communicate with for the unit tests.

There will be appsettings.json files in the test projects to tell the
glue code how to communicate with the SQL/ES dependency
services.  This makes things more cross-platform instead of
having a hard dependency on SQL LocalDb and ElasticsearchInside.