Have empty AppVeyor appsetting files
AppVeyor currently uses the build.ps1 to run the build,
and that build is spinning up Docker instances.  Therefore
since the appsetting overrides are in build.ps1, we do
not need to duplicate them here.

We could completely remove these files from git, but if 
we need them in the future we would have to remember
to set "CopyToOutputDirectory" to "Always" in the csproj.